Keep aligned vector in memory.hpp
