+++ include/sgl_kernel/jit_kernel/elementwise/qknorm.hpp
@@ -22,28 +22,19 @@
 #include <cstdint>
 #include <sycl/sycl.hpp>
 
+#include "../memory.hpp"
+
 namespace sgl {
 namespace sycl_kernel {
+
+// Import aligned_vector from memory utilities
+using sgl::sycl::aligned_vector;
 
 static constexpr int kQKNormSubGroupSize = 16;
 
 // ============================================================================
-// Vector Type Helpers
+// Vector Size Helper
 // ============================================================================
-
-template <typename T, int N>
-struct alignas(N * sizeof(T)) qknorm_aligned_vector {
-  T data[N];
-
-  qknorm_aligned_vector() = default;
-
-  T& operator[](int i) {
-    return data[i];
-  }
-  const T& operator[](int i) const {
-    return data[i];
-  }
-};
 
 template <typename T, int64_t kSize>
 constexpr int qknorm_get_vec_size() {
@@ -57,7 +48,7 @@ constexpr int qknorm_get_vec_size() {
 template <typename T, int64_t kHeadDim, int kVecSize = qknorm_get_vec_size<T, kHeadDim>()>
 class QKNormKernel {
  public:
-  using Vec = qknorm_aligned_vector<T, kVecSize>;
+  using Vec = aligned_vector<T, kVecSize>;
   static constexpr int64_t kVectorizedSize = kHeadDim / kVecSize;
 
   QKNormKernel(
